- Added GameObject class that holds a structure for an object. Two objects to use with the GameObject currently, they are the Mesh and Shape Object. It also holds an ObjectID which will be unique for each object created.
To create a GameObject, you have to utilize the ObjectManager within a scene to add an object. You just have to pass one of the two objects structs.

- Added Scene class that holds just a name and an object manager for now.

- Added Project class that holds the project name and a map of scenes within this particular project. More will certainly be added.

- Added Mesh structure which stores relevant data, such as a VkVertexData struct, texture name, VkDescriptor, color, position and scale. Much more will be added to this structure.

- Added Shape structure which stores relevant data, such as a texture name, color, position and scale. Much more will be added to this structure.

- Editor

You can now refresh the texture folder and it will aquire any new textures found. The textures of course will be added to the texture options and can be applied to objects.

Started working on the pre-application part of the editor, allowing a user to choose from an existing project or create a new one.

Had some issues with circular dependencies and have moved a few things around in other files and this file. This is still very much being thought about and I may restructure things in the future if it makes sense.